Separate duke into packages
