+++ duke.txt
@@ -1,4 +1,3 @@
 D,0,eat,2021-12-24
 T,1,return
-D,0,r,2001-10-10
 E,0,g,tdy,tmr
