--- src/main/java/Event.java
@@ -1,19 +1,0 @@
-public class Event extends Task {
-    protected String from;
-    protected String to;
-
-    public Event(String description, String from, String to) {
-        super(description);
-        this.from = from;
-        this.to = to;
-    }
-
-    @Override
-    public String toFileString() {
-        return "E," + super.toFileString() + "," + this.from + "," + this.to;
-    }
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + this.from + " to: " + this.to + ")";
-    }
-}
