--- src/main/java/Todo.java
@@ -1,16 +1,0 @@
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toFileString() {
-        return "T," + super.toFileString();
-    }
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-
-}
