+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 abstract public class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
     public abstract boolean isExit();
