+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class DeleteCommand extends Command {
     int taskNumber;
     public DeleteCommand(int taskNumber) {
@@ -6,7 +13,7 @@ public class DeleteCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         String t = tasks.delete(this.taskNumber);
-        ui.showResult("Ok, I've removed this task:");
+        ui.showResult("Ok, I've removed this duke.task:");
         ui.showResult(t);
         ui.showResult("Now you have " + tasks.size() + " task(s) in the list.");
     }
