+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class UnmarkCommand extends Command {
     int taskNumber;
     public UnmarkCommand(int taskNumber) {
@@ -7,7 +14,7 @@ public class UnmarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         String t = tasks.unmark(this.taskNumber);
-        ui.showResult("Nice, I've marked this task as done:");
+        ui.showResult("Nice, I've marked this duke.task as done:");
         ui.showResult(t);
     }
