+++ src/main/java/duke/task/Deadline.java
@@ -1,0 +1,29 @@
+package duke.task;
+
+import duke.DukeException;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public class Deadline extends Task {
+    protected LocalDate by;
+
+    public Deadline(String description, String by) throws DukeException {
+        super(description);
+        try {
+            this.by = LocalDate.parse(by);
+        } catch (DateTimeParseException e) {
+            throw new DukeException("Invalid date or date format");
+        }
+    }
+
+    @Override
+    public String toFileString() {
+        return "D," + super.toFileString() + "," + this.by;
+    }
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + this.by.format(DateTimeFormatter.ofPattern("d MMM yyyy")) + ")";
+    }
+}
