+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
 Hello! I'm Bob
 What can I do for you?
-Got it. I've added this task:
+Got it. I've added this duke.task:
 [T][ ] borrow book
-Now you have 1 task(s) in the list.
+Now you have 1 duke.task(s) in the list.
 Bye. Hope to see you again soon!
